+++ scripts/benchmark/motion_generation/artifacts.py
