+++ scripts/benchmark/motion_generation/registry.py
