+++ scripts/benchmark/motion_generation/scenarios/__init__.py
