+++ scripts/benchmark/motion_generation/scenarios/base.py
