+++ scripts/benchmark/motion_generation/scenarios/free_space.py
