+++ scripts/benchmark/motion_generation/suites/coverage.yaml
