+++ scripts/benchmark/motion_generation/suites/smoke.yaml
